requirements.txt
## Web Scraping Libraries
beautifulsoup4
requests
#DotEnv
python-dotenv